CHROMIUM: iwl7000: mvm: set different pm_timeout for action frames
When building a Tx Command for management frames, we are lacking
a check for action frames, for which we should set a different
pm_timeout.  This cause the fw to stay awake for 100TU after each
such frame is transmitted, resulting an excessive power consumption.

Signed-off-by: Avri Altman <avri.altman@intel.com>
iwl7000-tree: d6cd2436fdaa2fd5973d059d46f576c0695e1525