UPSTREAM: ASoC: Intel: Get correct usage_count value to load firmware
The usage_count variable was read before it was set to the correct
value, due to which the firmware load was failing. Because of this
IPC messages sent to the firmware were timing out causing a delay
of about 1 second while playing audio from the internal speakers.

With this patch the usage_count is read after the function call
pm_runtime_get_sync which will increment the usage_count variable
and the firmware load is successful and all the IPC messages are
processed correctly.

Signed-off-by: Shilpa Sreeramalu <shilpa.sreeramalu@intel.com>
Signed-off-by: Fang, Yang A <yang.a.fang@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Cc: stable@vger.kernel.org
(cherry picked from commit 412efa73dcd3bd03c1838c91e094533a95529039)

BUG=chrome-os-partner:42903
TEST=make sure no timeout err in the kernel log

Change-Id: I5fc386ddf5d7aecc0c7cf9f3dd0dc2adb708a6a1
Signed-off-by: Fang, Yang A <yang.a.fang@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/286543
Reviewed-by: Dylan Reid <dgreid@chromium.org>