UPSTREAM: arm64: kernel: Add optional CONFIG_ parameter to ALTERNATIVE()
Some uses of ALTERNATIVE() may depend on a feature that is disabled at
compile time by a Kconfig option. In this case the unused alternative
instructions waste space, and if the original instruction is a nop, it
wastes time and space.

This patch adds an optional 'config' option to ALTERNATIVE() and
alternative_insn that allows the compiler to remove both the original
and alternative instructions if the config option is not defined.

Suggested-by: Catalin Marinas <catalin.marinas@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit 91a5cefa2f98bdd3404c2fba57048c4fa225cc37)
Signed-off-by: Ricky Liang <jcliang@chromium.org>

BUG=chrome-os-partner:47644
TEST=Boot kernel on Oak.
TEST=Pass trybot run on smaug-release.

Change-Id: I13f0dd8a7188e273467223d94283a548766ab938
Reviewed-on: https://chromium-review.googlesource.com/312662
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>