CHROMIUM: drm/nouveau: Really fix sgt mappings that span multiple PDEs
Previous attempt fixed only single PDE overflow case, but with large
contiguous mapping it is possible that multiple overflows would occur
corrupting the mapping.

This patch attempts to fix the issue completely by restarting the loop
iteration with the same SG entry, but taking account for already mapped
part of it.

BUG=none
TEST=UI working fine

Change-Id: Id6268bf660a1b9a9070fc08ded3cac817905efcb
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/287666
Reviewed-by: Alexandre Courbot <gnurou@gmail.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>