CHROMIUM: iwl7000: pcie: don't synchronize_irq from the ISR
It is usually a bad idea to call synchronize_irq() from the
ISR. Even more so with a from the ISR and with mutex held.
Since rfkill is an interrupt and it can call
iwl_trans_stop_device, we can get to a situation where we
call synchronize_irq from the ISR with trans_pcie->mutex
held which prevents other threads from running.

For the moment, just remove the synchronize_irq in
iwl_trans_stop_device. We will need to find a solution for
the interrupts that might be waiting later.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 3ce0bde3bb97ecc9680eb157853c7f71090ab73d