drm/tegra: Wait for events to drain in commit
This patch alters the behavior of the atomic commit wait
such that if there is an event requested from a crtc state,
we wait for the event to be cleared instead of waiting one
vblank.

This fixes a bug where events could be lost if the frame is
not latched on the first vblank. In this case, the commit
routine will consider the frame finished and start processing
the next one, overwriting the old event in the process.

If there is no event requested, the behavior is unchanged.

Change-Id: I8d785980e10af8ebbeeecd86e49572a086ec2e32
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/289290
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>