CHROMIUM: usb: xhci-tegra: Fix up RESET_SSPI handling
A RESET_SSPI is needed whenever ID_OVERRIDE has been set to floating.
There are two situations in which this can happen: when switching to
device mode, and upon exiting LP0 without a cable connected.  In order
to account for the latter, always do a RESET_SSPI when switching to
host mode, regardless of whether or not we were previously in device
mode.

This also allows us to get rid of data_role_changed.  Remove it, and
add the initial call to tegra_xhci_update_data_role() to before the
firmware is loaded so that VBUS/ID overrides may be disabled, if
necessary, before the HC is started.  An initial RESET_SSPI is still
needed after the HC is started.

BUG=chrome-os-partner:43680
TEST=SuperSpeed device enumerates successfully after suspend/resume.

Change-Id: I663b8f43637dbcbd77bcf30d45dff37a1dc511e7
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/291447
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>