CHROMIUM: usb: gadget: tegra_xudc: Forcibly disconnect on suspend
Instead of simply aborting suspend if the device is connected,
force a disconnect by clearing the VBUS override and disabling
the controller.  Then, before powergating the controller, we must
wait for the disconnect event and for the interrupt handler to
copmlete.

This also fixes a race at cable disconnect where we may enter
runtime suspend before processing the disconnect event.

BUG=chrome-os-partner:43743
TEST=Suspend/resume with Smaug connected to host; adb still works
after resume.

Change-Id: I7903a514c9b8c5ae996602ac41dd61e0f7ddc46a
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/291840
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Tested-by: Mark Kuo <mkuo@nvidia.com>