CHROMIUM: drm/nouveau: gk104: channel timeout detection
Enable the scheduling timeout error interrupt and set it to a low value
to happen periodically, since it can be missed in HW in certain
conditions. Increment a channel-specific counter in software in the
error handler if the current channel hasn't advanced. Abort the channel
once the timeout limit is hit (with the periodic granularity). The error
notifier is set to NOUVEAU_GEM_CHANNEL_FIFO_ERROR_IDLE_TIMEOUT when this
occurs.

A new KEPLER_SET_CHANNEL_TIMEOUT mthd sets the timeout limit, in
milliseconds. The interrupt granularity is set to 100 ms. The default
limit is five seconds.

Some status bit mangling in the sched error handler is cleaned up too.

Change-Id: I3fb4543c830ff80de1beba4f4db8e345ed28a91a
Signed-off-by: Konsta Holtta <kholtta@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/292652
Reviewed-by: Haixia Shi <hshi@chromium.org>