CHROMIUM: dts: Set lp8557 backlight to pwm mode to enable cabc
This patch set lp8557 set control register of lp8557 backlight
to 0x02 to enable pwm mode to enable CABC. This has one drawback
that the brightness change will not be synchronous with the display.

We can set the register to 0 instead to enable synchromous change
but we need to modify the panel driver to change the brightness
there as current mechanism will be broken.

BUG=chrome-os-partner:43906
TEST=CABC works. Can change brightness using the same way.

Change-Id: I5985ba33995e80d698d185bc56965e8c670a7dba
Signed-off-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293563
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: John Kaehler <jkaehler@chromium.org>
Reviewed-by: Sean Paul <seanpaul@google.com>