CHROMIUM: smaug dts: add wifi wakeup-source (HACK)
Add wake event 8 (wifi_wake_ap) to the list of events that can
wake the AP.

This is a hack because we're adding the wake source to a
different device.  The wifi driver doesn't associate itself
with its entry in the device tree.

BUG=none
TEST=suspend & send a hangouts message from another device ->
wifi wakes the system

[ 3719.786465] PMC wake level = 0x10108000000100
[ 3719.790822] PMC wake enable = 0x18109001000108
[ 3719.795653] PMC wake status = 0x100
[ 3719.799163] Resume caused by PMC WAKE8, irq -6

BRANCH=none

Signed-off-by: Myles Watson <mylesgw@chromium.org>

Change-Id: I330802c4ab7bf7fd6717a48a28148a7f4f442109
Reviewed-on: https://chromium-review.googlesource.com/293862
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Myles Watson <mylesgw@chromium.org>
Tested-by: Myles Watson <mylesgw@chromium.org>